+++ src/main/java/gigi/commands/EventCommand.java
@@ -1,3 +1,5 @@
+package gigi.command;
+
 import java.time.LocalDateTime;
 
 public class EventCommand extends Command {
