+++ src/main/java/gigi/commands/ListCommand.java
@@ -1,3 +1,5 @@
+package gigi.command;
+
 public class ListCommand extends Command {
     public static final String COMMAND_WORD = "list";
