+++ src/main/java/gigi/commands/UnmarkCommand.java
@@ -1,3 +1,5 @@
+package gigi.command;
+
 public class UnmarkCommand extends Command {
     public static final String COMMAND_WORD = "unmark";
     public final int taskIndex;
