+++ src/main/java/gigi/storage/Storage.java
@@ -1,6 +1,9 @@
+import gigi.task.Deadlines;
+import gigi.task.Events;
+import gigi.task.Task;
+import gigi.task.ToDos;
+
 import java.io.*;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 
 public class Storage {
